Move allocations code into cas1 package
This logic only applies to CAS1
