+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/cas1/allocations/rules/EsapAssessmentRule.kt
@@ -1,12 +1,12 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.allocations.rules
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.allocations.rules
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.allocations.UserAllocatorRule
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.allocations.UserAllocatorRuleOutcome
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApprovedPremisesApplicationEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentEntity
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.allocations.UserAllocatorRule
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.allocations.UserAllocatorRuleOutcome
 
 @Component
 @ConditionalOnProperty(name = ["user-allocations.rules.esap-assessments.enabled"])
